django/requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.0.8
djangorestframework==3.12.2
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.16
psycopg2==2.8.6
pytz==2021.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
